+++ packages/iov-multichain/src/jsonrpcsigningserver.ts
@@ -1,5 +1,5 @@
 import { ChainId, isUnsignedTransaction, UnsignedTransaction } from "@iov/bcp";
-import { isJsonCompatibleDictionary } from "@iov/encoding";
+import { isJsonCompatibleDictionary, TransactionEncoder } from "@iov/encoding";
 import {
   jsonRpcCode,
   JsonRpcErrorResponse,
@@ -11,7 +11,6 @@ import {
 } from "@iov/jsonrpc";
 
 import { SigningServerCore } from "./signingservercore";
-import { TransactionEncoder } from "./transactionencoder";
 
 interface RpcCallGetIdentities {
   readonly name: "getIdentities";