+++ packages/iov-multichain/types/index.d.ts
@@ -1,4 +1,3 @@
 export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SignedAndPosted, SigningServerCore, } from "./signingservercore";
-export { TransactionEncoder } from "./transactionencoder";